discord-bot/pyproject.toml
[project]
name = "mealbot-discord-bot"
version = "1.0.0"
description = "Discord bot for mealbot that automatically extracts recipes from video URLs"
authors = [
    {name = "mealbot Team"}
]
requires-python = ">=3.8.1"
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]
